pyproject.toml
[project]
name = "photograph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dearpygui>=2.1.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
]
